Move domain module into separate file
